Rework question interface and add more questions
Make Question interface more testable and split into multiple extended
Question interfaces, i.e. RandomWordQuestion and WordListQuestion.

The Question interface have been changed so that the question functions
takes in the dynamic part of the question as an argument. A
randomX function is used to get a random value for the question.  The
correctAnswer function has been amended to take in both the player's
answer as well as the question parameters to check if the answer is
correct or not.

All these changes make the questions more testable. Instead of making
extended interfaces it is likely better to make a generic type to
specify the data type of the question parameters. Then the game
controller does not need to know the type of question to generate
questions, check, answers, etc.