feat: Additional questions and improved typing
The Question interface has been refactored to accept a generic type.
This generic type is the input that is returned from the random
function, passed as input to the question function, and the
correctAnswer function that checks if the answer is correct or not.

For any code that handles questions, each question object will work
exactly the same without having to deal with question types. A quiz
master routine can thus simply select questions from an array and don't
need to know about its implementation details.

All in all this commit brings the question count up to 16. Like the
original extreme_startup hackathon server this includes various word,
word list and number questions. Points for each question is a bit
arbitrary still, but more difficult questions reward more points.