feat: Add hint URLs to question objects
Optional hint URLs allows the game UI to show hints to players for the
more difficult questions. E.g. if you've never read about a caesar
cipher before an URL to the Wikipedia article will help you get going.

Renamed the isogram question to heterogram which seems to more precise
than isogram which can refer to both words with letter frequencies of 1,
and the wider case of words having the same frequency for every letter.