chore: Add more questions
Add binary to decimal, morse code decoder, and happy number questions.
Morse code decoder and happy number should be reasonably difficult and
nice to have for late game questions.

Binary to decimal can be solved relatively easily (especially with
Google and/or Copilot) so should not present too much of an obstacle.

With this commit I think we're nearing a good selection of questions.