feat: Initial game loop, scheduling, question picker
Introduce a game loop function that should be called for each player in
the game. The askPlayerQuestion function will recursively ask players
question using setTimeout to schedule each new question. By keeping
track of when questions are asked in the player log state we can
calculate any drift and adjust the set interval time of the next
question accordingly. This should reduce any game time drift caused by
the event loop blocking.

A rate limit function tries to keep track of the game answer trend by
looking at the last 20 log entries and checking if the score trends
upwards or downwards. For very good trends and negative trends we reduce
the rate of questions. For slightly positive trends we increase the rate
to allow players to climb more rapidly. The rate limit step is 1 second.

The admin dashboard has been improved UX wise. Not all functionality is
complete, but it is possible to start a demo game, stop the game, etc.