refactor: Redesign and improve player log state
This commit implements a slight redesign. Instead of using a background
image of a Mac behind a backdrop it now features a nice CSS generated
pattern. The pattern is muted enough not to be too distracting. It uses
fixed position to follow the user scrolling so it always covers the
background. This creates a nice subtle effect when the user scrolls.

A new game logo image has been generated using DALL·E. The image
features a neon futuristic cartoon feel depicting a collection of
monkeys ready to participate in the code monkey clash game. You can tell
it is AI generated as there are imperfections, but it is more than good
enough for a workshop game.

The PlayerLog state interface was updated to split out http status
codes and errors into separate properties instead of bundlign them up
into a union type property. Both variants would work, but distinct
properties are easier to use in the templating functions.

In order to avoid XSS and HTML injection from user answers the player
page now escapes html in the player log table. This way the player log
table does not simply display the HTML as is which I discovered when
entering the game server itself as a participant in the game.

Add a test client to use when running the game logic to get player with
positive points when doing testing.