chore: Web components for client side counters (#5)
Use client side logic in the form of a web component to count time.
Fetching time from server every x seconds is both bad UX and taxes the
server unecessarily. Instead we can ship a tiny bit of JS in the form of
a web component and just do the counting on the client.

This does not break the promise that Code Monkey Clash should be
perfectly runnable without JS, but the experience will be reduced in any
case as live updating via server rendering would require HTMX's
JavaScript client to run in any case.