chore: Handle each player on separate worker threads (#6)
Earlier implementation ran all the Code Monkey Clash server logic in a
single main thread. This includes the HTTP server and all rendering and
request logic, the game logic and player worker loop, as well as
persiting state. All this work adds up and can in the worst case lead to
unreliable messaging with contestant servers.

To help make the server more performant and the player game loop timings
more predictable logic has been split into a main worker, multiple
player game loop workers (one for each player), and finally a worker
used to persist game state.

The main thread runs first and sets up the Elysia server and is
responsible for handling HTTP requests, rendering HTML, etc. It
also handles co-ordination between the other workers.

The player worker threads handle the game loop for each player. It
stores player scores as well as some game info (round and state) to be
able to determine the question interval and which questions to pick
from for each round. The game loop is still implemented as setTimeout
recursively scheduling a new question. While timing is less of an issue
on a dedicated thread we still want to be as accurate as possible, so
drift is still accounted for.

The player worker thread receives events from the main thread like
game-started, player-joined, etc. Based on these it might set some
internal state, start or stop the game loop, or something else. Whenever
a player answers, or fails to answer, a question the worker will pass a
PlayerLog message to the main thread with the points, new score, error
message (if applicable), etc. The main worker then adds this to the
state.

A state persist worker is implemented to handle persisting state to
disk. In the old implementation state persitence could be triggered
concurrently by different functions and lead to corruption of the state
file. In the new implementation new state is passed to the state worker.
The worker then process one message at a time, writing state
synchronously to disk. This way no corruption should occur and state
writes happen on a first come, first served basis.