Add player rows and new scores to admin dashboard via SSE
When new players are added to state pass event to UI listeners and make
admin SSE render a PlayerRow to the SSE stream. This is then injected
into the table. Likewise update player scores (whenever we receive a
PlayerLog event) so that the score can be updated. We also recalculate
the total questions asked by the server and update that via the SSE
stream.