feat: Live admin dashboard via SSE (#9)
* feat: Live admin dashboard via SSE

Clean up SSE handling by introducing a helper function to set up a
ReadableStream object and associated logic to clean up the UI event
listener callback functions in the state.

Use new SSE helper function in the player and admin pages. This greatly
simplifies the logic in the sse endpoint handlers and makes them more
readable. Now it essentially boils down to declaring the callback
functions and ensure they only trigger for the events needed.

Clean up some logic and improve player worker.

* Add player rows and new scores to admin dashboard via SSE

When new players are added to state pass event to UI listeners and make
admin SSE render a PlayerRow to the SSE stream. This is then injected
into the table. Likewise update player scores (whenever we receive a
PlayerLog event) so that the score can be updated. We also recalculate
the total questions asked by the server and update that via the SSE
stream.