feat: Game scoreboard with live updates
Add new scoreboard page with live updating player scores and a chart.js
chart with the score progression for each player. Use colors to
differentiate each player so it is simpler to see how each player is
performing.

Uses sse channel as in admin and player pages to do live updates. For
the player list we simply send events keyed on the player uuid and
replace the number in the span for the score. Aditionally we re-sort the
list whenever there is a matching sse event.

For the chart we use chart.js by importing Chart.js, registering
necessary components, and exporting a renderChart(datasets) function in
the client.ts module. This can then be invoked from a <script> tag
embedded in the scoreboard page response.

Todo: Set up live updates of the chart.js data.