refactor: Questions, adjust simulation, fix rounds
Refactor question interface to make it simpler to run simulations. Each
question now needs to contain a match method which tells us if the
question matches some question string. It also needs to provide a solve
method which given the question string finds the answer. The base
question string is now a separate property to avoid duplication.

These changes makes it easy to run simulations as the question
implementations also provide the solution now. Each simulated player can
pick the solver functions from the question by searching the array for
the relevant question using the match function.

Adjusted simulated players

Each player in the simulation was too static. Instances of the expert
player would behave exactly the same. The example player logic was
adjusted to introduce some randomness and to make them more realistic.
For example, each player will not be able to answer a question before
having seen it a couple of times.

Adjust question rate intervals

The game was moving a bit slow during simulation, so question rate
minimum and maximum was adjusted. This should make the game feel a bit
quicker and allow players to mass points a bit more quickly. Might need
to adjust this again.

The buttons to change rounds only modified the state in the main worker.
None of the player workers were notified. This is now fixed, and was
necessary as this is used by the player workers to pick the correct
questions.