feat: Winner page when game is ended (#26)
Add a new page announcing the winners when game is over. The page
displays the three highest ranked player with medal emojis and score.
Shows confetti in the background to add some cheer to the game.

Add game status indicator to the scoreboard page with live updates using
SSE. When a game is stopped, paused or continued the status message
changes. In the case of the game starting the status message
disappears. When the status is game ended we navigate to the winner page
with the confetti.