fix: Player score not updated (#27)
When syncing player answer logs to main thread it only updated the
player log, not the player score. This broke UI that used the
player.score property rather than looking up the last score in the
player.log. Fixed game module so player score is updated correctly.
Changed some UI components to use the player.score property as this
should always be set.