chore: Optimize confetti implementation
To improve the performance of the confetti particle animation we can
move all data into a 64 bit float array to ensure data is stored
contiguously in memory so it is CPU cacheable and thus computations are
much faster.

With the new implementation I've increased the particle count to 5000.
It can handle more on newer hardware, but this seems sufficient. More
confetti would perhaps even be distracting.

Each confetti particle is implemented as a 10 parameter particle each
parameter a double precision float. Position 0 to 9 in the Float64Array
will then contain the first particle, position 10 to 19 the second, and
so forth.

We define offsets for each parameter so it is easier to extract the
relevant parameter values when we loop through the array. For example
the x offset is 7, so to get the x coordinate of the third particle is
as simple as doing: particleWidth * index + xOffset.

This implementation fixes the tilt logic and properly adds fluttering.